msm: avoid stack frame exceeding the limit of 1024 bytes
From: https://github.com/RaymanFX/kernel_sony_msm8x60/commit/f06c3ae83a63ddf2a047e270fcccbc1f15012ac8

The frame size of the {link|str_line} variable was exceeding the limit of
1024 bytes, hence GCC threw up a warning.
The cause of this was our stack frame maximum limit was nearly fulfilled
by other variables already, thus leaving no bytes for increasing the value
of an existing variable (which was caused by a CAF commit in this branch).
To move things out of the stack frame, we had two possibilities:
1) move the variable to heap memory
2) move the variable to global memory

Solution 1 would need a more complicated implementation, so it was decided
to go with the second solution, which uses the global memory.

Signed-off-by: Luis Cruz <ljc2491@gmail.com>

Conflicts:
	drivers/gpu/msm/adreno.c